pyproject.toml
[project]
name = "hydrostatic-balance"
version = "0.1.0"
description = ""
authors = [{ name = "Manuhhh", email = "manuelbadillabustelo@gmail.com" }]
readme = "README.md"
requires-python = "3.11.0"
dependencies = ["pasco (>=0.3.65,<0.4.0)", "serial (>=0.0.97,<0.0.98)", "python-dotenv (>=1.1.1,<2.0.0)"]

[tool.poetry]
packages = [{ include = "hydrostatic_balance", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.hydrostatic_balance.__init__:main"
